chore(backend): File refactor, adding all features folders
