--- backend/app/agents/__init__.py
@@ -1,11 +1,0 @@
-from app.agents.explore import run_explore_agent, run_explore_agent_streaming, OpenRouterClient
-from app.agents.graph import run_graph, run_graph_streaming, get_compiled_graph
-
-__all__ = [
-    "run_explore_agent",
-    "run_explore_agent_streaming",
-    "OpenRouterClient",
-    "run_graph",
-    "run_graph_streaming",
-    "get_compiled_graph",
-]
