+++ backend/app/explore/api/v1/router.py
@@ -1,5 +1,5 @@
 from fastapi import APIRouter
-from app.api.v1.endpoints import chat, documents
+from app.explore.api.v1.endpoints import chat, documents
 
 
 router = APIRouter(prefix="/v1")
