+++ backend/app/explore/schemas/document.py
