+++ backend/app/explore/services/pdf_parser.py
